clickup_tests/requirements.txt
playwright~=1.51.0
allure-python-commons~=2.14.1
pytest~=8.3.5
Faker~=37.1.0
requests~=2.32.3
python-dotenv~=1.1.0